move attachments related code to de.metas.attachments
https://github.com/metasfresh/metasfresh/issues/2074
