+++ de.metas.adempiere.adempiere/client/src/main/java-legacy/org/compiere/apps/Attachment.java
@@ -42,11 +42,8 @@ import org.adempiere.images.Images;
 import org.adempiere.pdf.Document;
 import org.adempiere.pdf.viewer.PDFViewerBean;
 import org.adempiere.plaf.AdempierePLAF;
-import org.adempiere.service.IAttachmentBL;
-import org.adempiere.service.IAttachmentDAO;
 import org.adempiere.util.Services;
 import org.adempiere.util.lang.impl.TableRecordReference;
-import org.compiere.model.AttachmentEntry;
 import org.compiere.model.I_AD_Attachment;
 import org.compiere.model.MSysConfig;
 import org.compiere.swing.CButton;
@@ -62,6 +59,9 @@ import com.google.common.io.Files;
 
 import de.metas.adempiere.Constants;
 import de.metas.adempiere.form.IClientUI;
+import de.metas.attachments.AttachmentEntry;
+import de.metas.attachments.IAttachmentBL;
+import de.metas.attachments.IAttachmentDAO;
 import de.metas.i18n.IMsgBL;
 import de.metas.i18n.Msg;
 import de.metas.logging.LogManager;
