+++ de.metas.payment.esr/src/test/java/de/metas/payment/esr/ESRImportTest.java
@@ -46,7 +46,6 @@ import org.adempiere.ad.wrapper.POJOWrapper;
 import org.adempiere.invoice.service.IInvoiceBL;
 import org.adempiere.invoice.service.IInvoiceDAO;
 import org.adempiere.model.InterfaceWrapperHelper;
-import org.adempiere.service.IAttachmentBL;
 import org.adempiere.service.IOrgDAO;
 import org.adempiere.service.ISysConfigBL;
 import org.adempiere.service.ISysConfigDAO;
@@ -68,6 +67,7 @@ import de.metas.adempiere.model.I_C_Currency;
 import de.metas.adempiere.model.I_C_Invoice;
 import de.metas.adempiere.service.IPeriodBL;
 import de.metas.allocation.api.IAllocationDAO;
+import de.metas.attachments.IAttachmentBL;
 import de.metas.document.IDocTypeDAO;
 import de.metas.document.engine.IDocActionBL;
 import de.metas.document.refid.api.IReferenceNoDAO;
